This is a complex way to fix the name ID across different systems
Basic idea is that find the unique characteristics for each of the features (total is 4). For example, it would have 2 dashes in SSN such as XXX-XX-XXXX and it could have one dash in EIN such as XXX-XXXXXXX. In my case, I find that there is some rules working for detect most of the not- matching case. This version will detect more unmatched values than easy-version.